# docs/openapi/v2.yaml
openapi: 3.1.0
...
info:
  title: OOHX API v2
  version: 0.1.0
  description: |
    Danh mục công khai của sàn OOHX. Phần này phục vụ trang công khai và
    ứng dụng Next.js.

    Không dùng /api/v1 cho việc mới: v1 là hợp đồng đang chạy với đối tác và
    giữ nguyên hành vi, kể cả những chỗ đặt tên sai (xem `price_per_slot_vnd`).
servers:
  - url: https://api.oohx.net
    description: Production
  - url: http://oohx.test
    description: Môi trường phát triển
...
paths:
...
  /api/v2/campaigns/{campaign}/payments:
    parameters:
      - { name: campaign, in: path, required: true, schema: { type: string } }
    get:
      tags: [payment]
      security: [{ sessionCookie: [] }]
      summary: Công nợ theo từng media owner, và các khoản đã tạo
      description: |
        **Công nợ tính theo từng owner, không phải theo tổng campaign.** Người
        mua chuyển thẳng cho từng media owner, nên `is_fully_paid` chỉ đúng khi
        **mọi** owner đã đủ — không phải khi tổng thu bằng tổng phải thu. Tổng
        tiền che khuất công nợ từng owner là F07 trong `FINDINGS.md`.

        Tiền bằng **VND số nguyên** ở mọi trường. VAT cộng một chỗ duy nhất
        (`PaymentService::withVat()`).

        **Không có thông tin ngân hàng ở đây.** `owner` chỉ có `id` và `name`;
        CLAUDE.md mục 2 cấm lộ `bank_*` và `billing_info`.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/PaymentOverview' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    post:
      tags: [payment]
      security: [{ sessionCookie: [] }]
      summary: Xác nhận một khoản chuyển khoản cho một media owner
      description: |
        **`amount` là đề nghị, không phải quyết định.** `PaymentService` quyết
        số tiền: không gửi thì nó lấy đúng công nợ còn lại của owner đó, gửi
        vượt công nợ thì trả 422 kèm số đúng, công nợ đã bằng 0 thì từ chối
        tạo thêm. Client không đặt được số tiền mình muốn (CLAUDE.md mục 5).

        Cần quyền `manage_payments` — **không phải** quyền xem. Vai trò
        `viewer` xem được campaign nhưng nhận 403 ở đây.

        Campaign phải ở trạng thái `approved` hoặc `active`.

        `payment_nonce` là mã chống trùng của **chính lần gửi này**, không phải
        token phiên: token phiên không đổi giữa các lần trả, nên trả một phần
        rồi quay lại trả nốt sẽ nhận lại khoản cũ đã hoàn tất và không tạo được
        khoản mới. Không gửi mã thì vẫn chạy — tầng service dùng lại khoản đang
        chờ của cùng owner.

        `method` cố tình chỉ có một giá trị: VNPay và MoMo chưa có đường chạy
        nào, và một enum ba giá trị trong khi chỉ một hoạt động là nói sai về
        năng lực của hệ thống (CLAUDE.md mục 8).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [method, accept_terms]
              properties:
                method:        { type: string, enum: [bank_transfer] }
                owner_slug:    { type: string, description: 'Dùng cái này. Phải là owner có màn hình trong campaign.' }
                owner_id:      { type: string, description: 'Khóa nội bộ — trang Blade dùng. Client ngoài nên gửi `owner_slug`.' }
                amount:        { type: number, minimum: 1000, nullable: true, description: Bỏ trống để trả đúng công nợ còn lại. }
                payment_nonce: { type: string, maxLength: 64, nullable: true }
                accept_terms:  { type: boolean }
      responses:
        '201':
          description: Đã tạo khoản chờ xác nhận
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    allOf:
                      - type: object
                        required: [payment]
                        properties:
                          payment: { $ref: '#/components/schemas/Payment' }
                      - $ref: '#/components/schemas/PaymentOverview'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: laravel_session
      description: |
        Sanctum dạng SPA. Client gọi `GET /sanctum/csrf-cookie` trước, rồi gửi
        kèm cookie phiên và header `X-XSRF-TOKEN` cho mọi yêu cầu ghi. Nhóm
        này **không** dùng token đối tác; token đối tác chỉ dùng cho `/api/v1`.
  schemas:
...
    Error:
      type: object
      description: Định dạng lỗi thống nhất cho mọi endpoint v2.
      required: [error, message, code, details]
      properties:
        error:   { type: string }
        message: { type: string }
        code:    { type: integer }
        details:
          type: array
          items:
            type: object
            properties:
              field:   { type: string }
              message: { type: string }
...
    Payment:
      type: object
      description: |
        Một khoản thanh toán. **Không có** `metadata`, `idempotency_key`,
        `gateway_ref`, `invoice_url` — xem `PaymentResource` để biết lý do từng
        trường. `owner` chỉ có `id` và `name`: số tài khoản không đi qua đây.
      properties:
        id: { type: string }
        owner:
          type: object
          nullable: true
          properties:
            id:   { type: string }
            name: { type: string }
        method:          { type: string, enum: [bank_transfer] }
        currency:        { type: string, enum: [VND] }
        amount:          { type: integer }
        status:          { type: string }
        transaction_ref: { type: string, nullable: true }
        invoice_number:  { type: string, nullable: true }
        due_date:        { type: string, format: date, nullable: true }
        paid_at:         { type: string, format: date-time, nullable: true }
        created_at:      { type: string, format: date-time, nullable: true }
    OwnerDebt:
      type: object
      description: Công nợ với MỘT media owner. Tiền bằng VND số nguyên.
      properties:
        owner:
          type: object
          properties:
            id:   { type: string }
            name: { type: string }
        cost:      { type: integer, description: Chưa gồm VAT. }
        vat:       { type: integer }
        total:     { type: integer, description: Đã gồm VAT — số phải trả cho owner này. }
        paid:      { type: integer, description: 'Đã trả, đã trừ phần phân bổ cho dòng đã hủy.' }
        refunded:  { type: integer }
        pending:   { type: integer, description: Khoản đã tạo nhưng chưa xác nhận. }
        remaining: { type: integer }
        is_paid:   { type: boolean }
    PaymentOverview:
      type: object
      required: [campaign, summary, by_owner, payments, can_pay]
      properties:
        campaign:
          type: object
          properties:
            id:     { type: string }
            code:   { type: string }
            status: { type: string }
        summary:
          type: object
          properties:
            currency:       { type: string, enum: [VND] }
            total_cost:     { type: integer }
            vat:            { type: integer }
            total_cost_vat: { type: integer }
            total_paid:     { type: integer }
            refunded:       { type: integer }
            pending:        { type: integer }
            remaining:      { type: integer }
            is_fully_paid:
              type: boolean
              description: |
                Đúng khi **mọi** owner đã nhận đủ, không phải khi tổng thu bằng
                tổng phải thu (F07).
        by_owner: { type: array, items: { $ref: '#/components/schemas/OwnerDebt' } }
        payments: { type: array, items: { $ref: '#/components/schemas/Payment' } }
        can_pay:  { type: boolean }
  responses:
    NotFound:
      description: Không tìm thấy
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationFailed:
      description: Dữ liệu gửi lên không hợp lệ
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TooManyRequests:
      description: |
        Vượt giới hạn tần suất. Phản hồi giữ nguyên `Retry-After` và các
        header hạn mức của `ThrottleRequests` — đổi định dạng body không được
        phá thông tin điều khiển của giao thức (Codex R41).
      headers:
        Retry-After:
          description: Số giây nên chờ trước khi gọi lại.
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: |
        Xem được nhưng không được làm. Khác 404: 404 nghĩa là bản ghi thuộc tổ
        chức khác và sự tồn tại của nó cũng là thông tin riêng; 403 nghĩa là
        bản ghi đang hiện trên màn hình người gọi, và thứ thiếu là quyền.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: |
        Chưa đăng nhập. Dùng **đúng** định dạng lỗi thống nhất, kể cả ở mã
        này — bên tiêu thụ không phải viết một nhánh đặc biệt cho 401.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...